Se modifica el servicio de crear factura para que se puedan incluir los productos

+++ spring-training2/src/main/java/mx/com/digitalchallengers/springtraining2/entidad/Producto.java
@@ -7,8 +7,6 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Data
@@ -29,7 +27,7 @@ public class Producto implements Serializable {
     private String nombre;
 
     @Column(name = "precio")
-    private int precio;
+    private float precio;
 
 
 }
